# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Perguntas de Conexão API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quiz/lead:
    post:
      operationId: createQuizLead
      tags: [quiz]
      summary: Save a quiz lead and send its diagnosis email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizLeadInput"
      responses:
        "201":
          description: Quiz lead recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizLeadResponse"
        "400":
          description: Invalid quiz lead
        "429":
          description: Too many requests
  /email/sair/{id}:
    post:
      operationId: optOutOfEmail
      tags: [email]
      summary: Opt out of future checkout emails
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 120 }
      responses:
        "200":
          description: Opt-out accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailOptOutResponse"
...
components:
  schemas:
...
    QuizLeadInput:
      type: object
      properties:
        visitorKey: { type: string, minLength: 1, maxLength: 120 }
        lpId: { type: string, enum: [v1, v2, lp3] }
        email:
          type: string
          maxLength: 254
          pattern: "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$"
        diagnosisLabel: { type: string, minLength: 1, maxLength: 120 }
        diagnosisCopy: { type: string, minLength: 1, maxLength: 4000 }
        utmSource: { type: string, maxLength: 160 }
        utmMedium: { type: string, maxLength: 160 }
        utmCampaign: { type: string, maxLength: 200 }
        utmContent: { type: string, maxLength: 200 }
        utmTerm: { type: string, maxLength: 200 }
      required:
        [visitorKey, lpId, email, diagnosisLabel, diagnosisCopy]
    QuizLeadResponse:
      type: object
      properties:
        leadId: { type: string }
      required: [leadId]
    EmailOptOutResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...