# lib/api-zod/src/generated/types/quizLeadInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Perguntas de Conexão API
 * OpenAPI spec version: 0.1.0
 */
import type { QuizLeadInputLpId } from './quizLeadInputLpId';

export interface QuizLeadInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  visitorKey: string;
  lpId: QuizLeadInputLpId;
  /**
     * @maxLength 254
     * @pattern ^[^@\s]+@[^@\s]+\.[^@\s]+$
     */
  email: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  diagnosisLabel: string;
  /**
     * @minLength 1
     * @maxLength 4000
     */
  diagnosisCopy: string;
  /** @maxLength 160 */
  utmSource?: string;
  /** @maxLength 160 */
  utmMedium?: string;
  /** @maxLength 200 */
  utmCampaign?: string;
  /** @maxLength 200 */
  utmContent?: string;
  /** @maxLength 200 */
  utmTerm?: string;
}